Extracted VisualToolStripMenuItem to file
