+++ VisualPlus/VisualPlus.csproj
@@ -207,6 +207,9 @@
     <Compile Include="Toolkit\Controls\VisualKnob.cs">
       <SubType>UserControl</SubType>
     </Compile>
+    <Compile Include="Toolkit\VisualBase\VisualToolStripMenuItem.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Service Include="{94E38DFF-614B-4cbd-B67C-F211BB35CE8B}" />
   </ItemGroup>
   <ItemGroup>
